Restructure of react file structure
Just some basic restructuring of the react portions structure.
