+++ smartcar_frontend/src/components/difficulty/Difficulty.js
