+++ smartcar_frontend/src/components/difficulty/DifficultyOptions.css
