+++ smartcar_frontend/src/components/leaderboard/LeaderBoard.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import "./Home.css";
+import "../../pages/home/Home.css";
 
 //const LeaderBoardList = "";
