+++ smartcar_frontend/src/components/race/Race.css
