+++ smartcar_frontend/src/pages/home/Home.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { NavLink } from "react-router-dom";
-import logo from "../logo.png";
+import logo from "../../resources/logo.png";
 import "./Home.css";
-import Difficulty from "./Difficulty";
+import Difficulty from "../../components/difficulty/Difficulty";
 
 class HomeComponent extends React.Component {
   state = { difficulty: false, difficultyLevel: "Easy" };
