+++ smartcar_frontend/src/resources/logo.png
